refactor(layout): Sidebar sous 300 lignes — éclatement complet (H12)
- useSidebar.js: état repli/sous-menus, infos user, menu (buildMenuSections),
  persistance scopée institution, ouverture auto du sous-menu actif
- SidebarHeader/SidebarNav/SidebarUserProfile: sections présentationnelles
  + CSS déplacé verbatim (transitions fade dupliquées par composant, scoped)
- Sidebar.vue: orchestrateur 77 lignes (536 → 77)
- buildMenuSections (utils/sidebarMenu) déjà extrait : consommé, non touché
- Tests: useSidebar (logique) + montage des 3 sous-composants; Sidebar.test conservé
